Exclude python 3.4 since it is not supported

+++ ci/travis/manylinux.sh
@@ -6,28 +6,31 @@ set -e -x
 # Create wheels
 for PYBIN in /opt/python/*/bin; do
 
-	# Install and link cmake
-	${PYBIN}/pip install cmake
-	rm -rf /usr/local/bin/cmake
-	ln -s $PYBIN/cmake /usr/local/bin/cmake
+	if [[ ${PYBIN} != *"34"* ]]; then
 
-	# Install python dependencies
-	${PYBIN}/pip install numpy scipy future pytest
+		# Install and link cmake
+		${PYBIN}/pip install cmake
+		rm -rf /usr/local/bin/cmake
+		ln -s $PYBIN/cmake /usr/local/bin/cmake
 
-	# Create wheel
-	${PYBIN}/pip wheel /io/ -w dist/
+		# Install python dependencies
+		${PYBIN}/pip install numpy scipy future pytest
 
-        # Bundle external shared libraries into the wheels
-	for whl in dist/osqp*.whl; do
-	    auditwheel repair "$whl" -w /io/dist/
-	done
+		# Create wheel
+		${PYBIN}/pip wheel /io/ -w dist/
 
-	# Install package and test
-	${PYBIN}/pip install osqp --no-index -f /io/dist
+		# Bundle external shared libraries into the wheels
+		for whl in dist/osqp*.whl; do
+			auditwheel repair "$whl" -w /io/dist/
+		done
 
-	# Test
-	cd /io/
-	${PYBIN}/pytest
+		# Install package and test
+		${PYBIN}/pip install osqp --no-index -f /io/dist
+
+		# Test
+		cd /io/
+		${PYBIN}/pytest
+	fi
 
 done
